Exception Handling

+++ db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from db.hash import Hash
 from sqlalchemy.orm.session import Session
 from schemas import UserBase
@@ -22,11 +23,19 @@ def get_all_users(db: Session):
 
 
 def get_user_by_id(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found"
+        )
 
 
 def update_user(db: Session, id: int, request: UserBase):
-    user = db.query(DbUser).filter(DbUser.id == id)
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found"
+        )
     user.update(
         {
             DbUser.username: request.username,
@@ -40,6 +49,10 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found"
+        )
     db.delete(user)
     db.commit()
     return "Deleted"
